DamageEvent 정리
- floating text 출력
- 결과창 임시 생성
- camera shake 추가

+++ Game/Battle/Cutscene/VisualEffectEvent.cpp
@@ -25,7 +25,7 @@ bool VisualEffectEvent::Update(Wannabe::BattleContext& context, float fDeltaTime
 
     CanvasCell clearCell;
     clearCell.wch = L' ';
-    renderSys.GetEffectCanvas().SetCell((int)m_LastPos.x, (int)m_LastPos.y, clearCell);
+    renderSys.GetEffectCanvas().SetCell(m_LastPos.x, m_LastPos.y, clearCell);
 
     m_LastPos = m_CurrentPos;
 
@@ -33,17 +33,17 @@ bool VisualEffectEvent::Update(Wannabe::BattleContext& context, float fDeltaTime
     if (m_eMoveType == EffectMovementType::Straight)
     {
         // 선형 보간 (Lerp)
-        m_CurrentPos.x = m_StartPos.x + (m_EndPos.x - m_StartPos.x) * t;
-        m_CurrentPos.y = m_StartPos.y + (m_EndPos.y - m_StartPos.y) * t;
+        m_CurrentPos.x = static_cast<int>(m_StartPos.x + (m_EndPos.x - m_StartPos.x) * t);
+        m_CurrentPos.y = static_cast<int>(m_StartPos.y + (m_EndPos.y - m_StartPos.y) * t);
     }
     else if (m_eMoveType == EffectMovementType::Parabola)
     {
         // 포물선: X는 선형, Y는 2차 함수 (위로 솟구쳤다 떨어짐)
-        m_CurrentPos.x = m_StartPos.x + (m_EndPos.x - m_StartPos.x) * t;
+        m_CurrentPos.x = static_cast<int>(m_StartPos.x + (m_EndPos.x - m_StartPos.x) * t);
 
         float height = 5.0f; // 최대 높이 값
         float jump = 4.0f * height * t * (1.0f - t); // 2차 곡선 공식
-        m_CurrentPos.y = (m_StartPos.y + (m_EndPos.y - m_StartPos.y) * t) - jump;
+        m_CurrentPos.y = static_cast<int>((m_StartPos.y + (m_EndPos.y - m_StartPos.y) * t) - jump);
     }
 
     // 3. 새로운 위치에 그리기
@@ -52,7 +52,7 @@ bool VisualEffectEvent::Update(Wannabe::BattleContext& context, float fDeltaTime
     newCell.color = m_Color;
     newCell.sortingOrder = 100;
     newCell.bDirty = true;
-    renderSys.GetEffectCanvas().SetCell((int)m_CurrentPos.x, (int)m_CurrentPos.y, newCell);
+    renderSys.GetEffectCanvas().SetCell(m_CurrentPos.x, m_CurrentPos.y, newCell);
 
     return m_fElapsed >= m_fDuration;
 }
@@ -61,7 +61,7 @@ void VisualEffectEvent::OnEnd(Wannabe::BattleContext& context)
 {
     CanvasCell cell;
     cell.wch = L' ';
-    Engine::Get().GetRenderSystem().GetEffectCanvas().SetCell((int)m_CurrentPos.x, (int)m_CurrentPos.y,cell);
+    Engine::Get().GetRenderSystem().GetEffectCanvas().SetCell(m_CurrentPos.x, m_CurrentPos.y,cell);
 }
 
 bool VisualEffectEvent::IsValid(Wannabe::BattleContext& context) const
